+++ CHANGELOG.md
@@ -29,14 +29,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Improve resource card layout [#65](https://github.com/IN-CORE/incore-studio/issues/65)
-<<<<<<< HEAD
 - Improve performance by switching from Webpack to ESBuild [#200](https://github.com/IN-CORE/incore-studio/issues/200)
-=======
 - Disable unused buttons from navbar [#215](https://github.com/IN-CORE/incore-studio/issues/215)
+- Replace map preview with geo-explorer [#218](https://github.com/IN-CORE/incore-studio/issues/218)
 - Project Dashboard look and feel [#216](https://github.com/IN-CORE/incore-studio/issues/216)
 - Execution Page UI and Analysis Node Look and feel [#223](https://github.com/IN-CORE/incore-studio/issues/223)
 - UI improvements for Workflow Page [#217](https://github.com/IN-CORE/incore-studio/issues/217)
->>>>>>> develop
 
 ## [Beta-1] - 05-01-2025